# Avoid duplicate host load in allowed_hosts_obj