@simaishi 2018-06-21
Fine backport details:

```
$ git log -1
commit 361707d59833a4b0b388e4f880b6e7cc9f022c84
Author: Greg McCullough <gmccullo@redhat.com>
Date:   Tue May 22 15:45:25 2018 -0400

    Merge pull request #17402 from NickLaMuro/lighter_queries_in_miq_request_workflow
    
    Avoid duplicate host load in allowed_hosts_obj
    (cherry picked from commit d577df7a6d6f28c24264a65b30d7011db1d6d71a)
    
    https://bugzilla.redhat.com/show_bug.cgi?id=1593798
```