@simaishi 2018-06-21
Gaprindashvili backport details:

```
$ git log -1
commit 6296c6044b096edc6d45bd83017f504163e9f785
Author: Greg McCullough <gmccullo@redhat.com>
Date:   Tue May 22 15:45:25 2018 -0400

    Merge pull request #17402 from NickLaMuro/lighter_queries_in_miq_request_workflow
    
    Avoid duplicate host load in allowed_hosts_obj
    (cherry picked from commit d577df7a6d6f28c24264a65b30d7011db1d6d71a)
    
    https://bugzilla.redhat.com/show_bug.cgi?id=1593797
```